+++ packages/win32/CHANGELOG.md
@@ -218,6 +218,8 @@ All notable changes to this project will be documented in this file.
 - Deprecate `isCOMInitialized` getter in favor of `isComInitialized` (#903)
 - You can automatically migrate your code to use the new constants and
   properties by running `dart fix --apply` in your terminal.
+- Add `IsPropVariantString`, `PropVariantToStrRet`, and
+  `PropVariantToStringWithDefault` APIs (#904)
 
 ## [5.5.5] - 2024-10-02